# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rusting Frog API
  description: |
    Screaming-Frog-compatible SEO crawler backend, in Rust.

    **Auth.** Every endpoint requires a Laravel-issued JWT in `Authorization: Bearer <jwt>`.
    Dev: mint one with `node scripts/mint_jwt.js` (60-minute TTL; claims
    `tenant_id`, `user_id`, `scopes`).

    **Tenancy.** `tenant_id` is taken from the JWT `tenant_id` claim; callers never
    pass it. All reads and writes are scoped to the caller's tenant.

    **Implementation progress.** Batches 1–9 of the evaluator roll-out are wired:
      - *Batch 1 — response/title/meta/H1/H2* (19 filter keys): `response_code_*`,
        `internal_*`, `external_*`, `title_*`, `meta_descripton_*`,
        `meta_keywords_*`, `h1_*`, `h2_*`.
      - *Batch 2 — images / links / external* (32 filter keys): `images_*`,
        `links_*`, full `external_*` split by content-type.
      - *Batch 3 — canonicals / pagination / directives* (34 filter keys):
        `canonicals_*` (self-ref, canonicalised, missing, multiple,
        conflicting, relative, fragment, invalid, outside-head),
        `pagination_*` (contains, first-page, paginated-2-plus,
        not-in-anchor, multiple), `directives_*` (every robots token —
        index/noindex/follow/nofollow/none/noarchive/nosnippet/
        max-snippet/max-image-preview/max-video-preview/noodp/noydir/
        noimageindex/notranslate/unavailable-after/refresh/outside-head).
...
        via Flesch–Kincaid).
...
        flagged),
...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/projects/{project_id}/crawls:
    parameters:
      - in: path
        name: project_id
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [crawls]
      summary: List crawls for a project
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CrawlSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [crawls]
      summary: Start a crawl (async — returns immediately with status=queued)
      description: |
        Enqueues a crawl for this project. The HTTP response returns
        immediately with the new `crawl_id`; actual fetching runs in the
        `sf-crawl-worker` binary. Poll `GET /v1/crawls/{id}` for progress,
        or subscribe to the future WebSocket stream at `/v1/crawls/{id}/live`.

        The crawl honours the project's config snapshot at start time
        (max depth, respect-robots, render flag, etc.). Each URL is
        fetched, parsed, and evaluated through all implemented evaluator
        tabs (Batch 1 + 2 at present). Findings land in `crawl_url_findings`.
      responses:
        "200":
          description: Crawl queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrawlStartResponse" }
              examples:
                boltrack:
                  value:
                    id: 019d99c6-80f0-7251-a548-756acad321d4
                    status: queued
                    message: crawl queued
        "404":
          description: Project not found
          $ref: "#/components/responses/NotFound"
  /v1/crawls/{id}:
    parameters:
      - { $ref: "#/components/parameters/CrawlId" }
    get:
      tags: [crawls]
      summary: Get crawl status + counters
      description: |
        Returns crawl metadata and lifecycle state. Use this to poll a
        running crawl — `urls_discovered` grows as the frontier expands
        and `urls_crawled` grows as fetches complete. When `status =
        "completed"`, `completed_at` is set and no more URLs will be added.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrawlDetail" }
              examples:
                boltrack_complete:
                  summary: Completed boltrack.in crawl (210 URLs, ~110s)
                  value:
                    id: 019d99c6-80f0-7251-a548-756acad321d4
                    project_id: 019d99c6-5f95-7c72-b9a0-de12892deeb5
                    status: completed
                    urls_discovered: 213
                    urls_crawled: 210
                    seed_urls: ["https://boltrack.in/"]
                    created_at: 2026-04-17T04:50:16.432606Z
                    started_at: 2026-04-17T04:50:17.751117Z
                    completed_at: 2026-04-17T04:52:06.991479Z
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/crawls/{id}/pause:
    parameters:
      - { $ref: "#/components/parameters/CrawlId" }
    post:
      tags: [crawls]
      summary: Pause a running crawl (idempotent)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrawlStartResponse" }
  /v1/crawls/{id}/resume:
    parameters:
      - { $ref: "#/components/parameters/CrawlId" }
    post:
      tags: [crawls]
      summary: Resume a paused crawl
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrawlStartResponse" }
  /v1/crawls/{id}/stop:
    parameters:
      - { $ref: "#/components/parameters/CrawlId" }
    post:
      tags: [crawls]
      summary: Stop a crawl and finalize
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CrawlStartResponse" }
  /v1/crawls/{id}/robots:
    parameters:
      - { $ref: "#/components/parameters/CrawlId" }
    get:
      tags: [crawls]
      summary: Get the robots.txt body + parsed groups used for this crawl
      description: |
        Returns the literal `robots.txt` body fetched at crawl start, the
        HTTP status of that fetch, a best-effort structural parse (grouped
        user-agent → disallow/allow/sitemap) and the running count of URLs
        that were discovered but blocked by a `Disallow` directive.

        Status semantics:
          - `200` — site served a robots file; `raw` is populated and was
            used to gate every fetch.
          - `404` — site has no robots file; `raw` is `null` and every URL
            was allowed (matches Screaming Frog's default).
          - `null` — robots fetch errored (DNS / timeout / TLS); `raw` is
            `null` and the crawler fetched everything.

        `blocked_url_count` drives SF's "Response Codes → Blocked by
        Robots.txt" filter.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [url, status, raw, groups, blocked_url_count]
                properties:
                  url:
                    type: string
                    format: uri
                    nullable: true
                    description: The URL the robots.txt was fetched from.
                  status:
                    type: integer
                    nullable: true
                    description: HTTP status of the robots fetch. `null` means the fetch errored.
                  raw:
                    type: string
                    nullable: true
                    description: Literal robots.txt body. `null` when the fetch was non-2xx or errored.
                  groups:
                    type: array
                    description: One entry per `User-agent:` block, parsed out of `raw`.
                    items:
                      type: object
                      required: [user_agents, disallow, allow, sitemap]
                      properties:
                        user_agents: { type: array, items: { type: string } }
                        disallow:    { type: array, items: { type: string } }
                        allow:       { type: array, items: { type: string } }
                        sitemap:     { type: array, items: { type: string } }
                  blocked_url_count:
                    type: integer
                    description: Number of URLs in this crawl that were discovered but blocked by a matching `Disallow` directive.
              example:
                url: "https://github.com/robots.txt"
                status: 200
                raw: "User-agent: *\nDisallow: /*/pulse\nDisallow: /*/forks\nSitemap: https://github.com/sitemap.xml\n"
                groups:
                  - user_agents: ["*"]
                    disallow: ["/*/pulse", "/*/forks"]
                    allow: []
                    sitemap: ["https://github.com/sitemap.xml"]
                blocked_url_count: 2
  /v1/crawls/{id}/sitemap:
    parameters:
      - { $ref: "#/components/parameters/CrawlId" }
    get:
      tags: [crawls]
      summary: Get the sitemap.xml body + URL set used for this crawl
      description: |
        Returns the sitemap.xml body fetched at crawl start (or each of the
        child sitemap bodies if the root was a `<sitemapindex>`), the HTTP
        status of that fetch, and the parsed URL set with coverage stats.

        Coverage semantics:
          - `url_count` — total unique `<loc>` URLs extracted.
          - `crawled_count` — how many of those were actually visited by
            the crawler.
          - `orphan_count` = `url_count − crawled_count`; drives the SF
            "Orphan URLs" filter (URLs in sitemap but never discovered).

        The `urls` array is paginated via `?limit=N` (default 100, max
        1000); the totals always reflect the full set.
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
          description: Max URLs to return in the `urls` array.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [url, status, raw, url_count, crawled_count, orphan_count, urls, urls_returned]
                properties:
                  url:
                    type: string
                    format: uri
                    nullable: true
                    description: The URL the sitemap.xml was fetched from.
                  status:
                    type: integer
                    nullable: true
                    description: HTTP status of the sitemap fetch. `null` means the fetch errored.
                  raw:
                    type: string
                    nullable: true
                    description: Literal sitemap body (the root document only; child sitemap bodies are not stored).
                  url_count:
                    type: integer
                    description: Total unique URLs extracted from the sitemap(s).
                  crawled_count:
                    type: integer
                    description: Count of sitemap URLs also present in the crawl.
                  orphan_count:
                    type: integer
                    description: Count of sitemap URLs NOT visited by the crawler.
                  urls_returned:
                    type: integer
                    description: Size of the `urls` array in this response (honours `?limit=`).
                  urls:
                    type: array
                    items: { type: string, format: uri }
              example:
                url: "https://www.cloudflare.com/sitemap.xml"
                status: 200
                raw: "<?xml version=\"1.0\" encoding=\"UTF-8\"?><urlset xmlns=\"http://www.sitemaps.org/schemas/sitemap/0.9\">…</urlset>"
                url_count: 1820
                crawled_count: 1
                orphan_count: 1819
                urls_returned: 100
                urls:
                  - "https://www.cloudflare.com/"
                  - "https://www.cloudflare.com/about-overview/"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived JWT issued by Laravel; `tenant_id`, `user_id`, `scopes` claims required.
  parameters:
...
    CrawlId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found in this tenant
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
      required: [title, status]
...
    CrawlSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, paused, completed, failed] }
        urls_discovered: { type: integer }
        urls_crawled: { type: integer }
        started_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
      required: [id, status, urls_discovered, urls_crawled, created_at]
    CrawlDetail:
      allOf:
        - $ref: "#/components/schemas/CrawlSummary"
        - type: object
          properties:
            project_id: { type: string, format: uuid }
            seed_urls:
              type: array
              items: { type: string, format: uri }
    CrawlStartResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string }
        message: { type: string }
      required: [id, status, message]
...